refactor(client): rename add and delete tracks tests and use global utils
